backend/requirements.txt
# Framework principal
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0

# Banco de dados e autenticação
supabase==1.0.3
python-multipart==0.0.6

# Processamento de arquivos
pymupdf==1.23.8
python-docx==0.8.11
pytesseract==0.3.10
pillow==10.0.1

# APIs de IA
openai==1.3.0
requests==2.31.0
langdetect==1.0.9

# Processamento de áudio (opcional)
pydub==0.25.1
SpeechRecognition==3.10.0

# Utilitários
uuid==1.30
python-dateutil==2.8.2
pyyaml==6.0.1

# Docker para code executor (opcional)
docker==6.1.3

# Segurança
bcrypt==4.0.1
cryptography==41.0.4

# Desenvolvimento (apenas para dev)
blinker==1.6.3
debugpy==1.6.7

# Adicione estas linhas:
numpy==1.26.4
scipy==1.11.4
plotly==5.24.1
kaleido==0.2.1
matplotlib==3.8.0
trimesh==4.4.9
numpy-stl==3.1.1
